% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plik.R
\name{insertion_sort}
\alias{insertion_sort}
\title{InsertionSort}
\usage{
insertion_sort(x)
}
\arguments{
\item{x}{numeric vector to be sorted}
}
\value{
Returns a sorted numeric vector.
}
\description{
Sorts a numeric vector with the InsertionSort algorithm
}
\examples{
insertion_sort(c(4, 2, 8, 1, 5))
stopifnot(insertion_sort(c(3, 1, 2)) == c(1, 2, 3))
}
