% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plik.R
\name{quick_sort}
\alias{quick_sort}
\title{QuickSort}
\usage{
quick_sort(x)
}
\arguments{
\item{x}{numeric vector to be sorted}
}
\value{
Returns a sorted numeric vector.
}
\description{
Sorts a numeric vector with the Quick Sort algorithm.
}
\examples{
quick_sort(c(4, 2, 8, 1, 5))
stopifnot(quick_sort(c(3, 1, 2)) == c(1, 2, 3))
}
