% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plik.R
\docType{package}
\name{pakietR-package}
\alias{pakietR}
\alias{pakietR-package}
\title{pakietR: Sorting algorithms}
\description{
selection sort, insertion sort, quick sort
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gdrobot.github.io/pakietR/}
}

}
\author{
\strong{Maintainer}: Wojciech Florek \email{01199223@pw.edu.pl}

Authors:
\itemize{
  \item Wojciech Florek \email{01199223@pw.edu.pl}
}

}
